Allow functions to be passed into attribute check tuples
Now allows functions to run checks with 2-tuples in addition to having
lists of valid values.  Checks using this functionality should use
avoid using functions starting with `check` for the time being and
should return a `ratable_result`, which returns an anonymous function
with the `weight` parameter of the `Result` class unbound.  By putting
the 2-tuple with function in the appropriate check list, the proper
weight will automatically be assigned.  This has a number of advantages:
It's easier to check required attributes against a specification since
the first attribute of the 2-tuple is the attribute name.  Also, if an
attribute changes its weighting/importance in another revision of a
standard, it can be trivially changed simply by including the 2-tuple
with attribute name and function in the proper list for that particular
standard version.

Also renames `geospatial_bounds_valid_wkt` to `geospatial_bounds` inside
function return results.