Adds basic support for command line cchecker versions
Allows specification of a version on certain checkers from the command
line, i.e. `compliance-checker -t acdd:1.3'`.  For versioned checkers,
the default checker if no version is specified is the latest version.
Thus if both 'acdd:1.1' and 'acdd:1.3' exist, the checker will use the
'acdd:1.3' version if 'acdd' is specified on the command line.
Addresses compliance checker metadata mentioned in #186.