Fixes _find_boundary_vars
_find_boundary_vars now returns a list of strings. My plan is to
ultimately remove this method in favor of cfutil.get_cell_boundary_map
or cfutil.get_cell_boundary_variables.

The previous version took advantage of a cache but sacrificed
consistency and clarity for it. Iterating through a datasets variables
without accessing data is virtually trivial and the cache isn't needed
in my opinion.