Adds better support for detecting dataset protocol
We recently found out that Hyrax and ERDDAP did not work because the
server intitally returns a HTTP 200/Content-type: text/html when an HTTP
request is made to a DODS endpoint. This is so user's don't see the
error page that THREDDS shows when a user visits a pure DODS endpoint.

We relied on this behavior of THREDDS and the too literal interpretation
of DAPv2's specification to identify services implementing OPeNDAP.

This patch includes better identification of services implementing
OPeNDAP, and SOS. As well as a faster check for binary datasets like
netCDF.

Identification now works by checking if 'xdods-server' exists in the
response headers. In the case for SOS, if the response headers has
'text/xml' in the content-type header, then it is treated as an SOS
response.

Binary netCDF files are checked by the filename suffix or the magic
number (first four bytes of a file).