Mock NetCDF variable object optionally can copy attributes
Modified the constructor of the MockVariable class to optionally accept
a NetCDF variable.  When passed in, the mock variable object will
receive all the attributes from the variable.  This assists in testing
slight permutations of NetCDF files which were opened in read-only mode
without needing to rewrite the file, load it into memory, etc.