Checks formula_terms variables and well-formedness
Checks that the variables referenced in formula_terms attribute for
dimensionless vertical coordinates actually exist in the dataset prior
to referencing the formula from CF Appendix D.  Previously, if the
standard_name was not in Appendix D, the variables would not be checked.
This checks regardless of whether the standard_name actually exists in
Appendix D and whether or not the formula terms correspond to those in
Appendix D.  Also checks whether the formula_terms attribute is properly
syntatically formed.  Fixes #505.