Avoid crashing when finding "special variables" for CF
Avoids unhandled exceptions previously caused by attempting to call
lower() on various attributes when said attributes were not a string
type.  For the sake of finding variables, this code attempts to coerce the non
string attributes to string warning the user.  If it fails, another
warning is raised to indicate that the variable will be ignored because
it is not castable. (not sure if this is reachable with supported NetCDF
 dataset types, as most {all?} can cast to string)