CF Integration Tests Fixes
In order for the CF Integration Tests to pass, they needed some updating. Many test methods had assertion phrasing
that needed to be reworded to match the current check version. Instead of hard-coding the std. names table
version into the assertions, the code now gets it for us (this was one of the most common failures).

One failure case I am unsure how to approach is in `test_glcfs`, l. 252, where as assertion states that the boundary
variable `eta` does not exist in the dataset. While this is true, the message is never generated by the CF check
because the dataset has no variables with the `bounds` attribute, and thus does not test the variables referenced in
the `formula_terms` attribute. Any guidance on this would be appreciated.