Verifying Special-Attribute Data Types
Per issue #468, some special attributes defined in the NetCDF User Guide (NUG)
had data types which did not match their parent variable. E.g.
    float temperature(time);
        temperature:standard_name = "air_temperature";
        temperature:units = "deg_C";
        temperature:valid_range = -20, 50; # --> integers, not floats

Currently, type-mismatches for these special attributes are checked in
"check_packed_data()" for only variables which are packed. This commit
adds a new check which examines each variable. If the variable has one
of these special attributes and a type-mismatch exists, it is scored.
Unit tests for the new check are included.

Additionally, there was a global variable "__stdname_table__" at the top
of "cf.py"; as this is not used anywhere else, this commit removes it.

It is worth noting that, while "actual_range", "actual_min", and
"actual_max" are not defined in the CF-1.6 spec, this check includes
them for easier integration into the CF-1.7.