Allows checkers using name .attr to be listed
Allows checkers using the `.name` attribute instead of the `._cc_spec` and
`cc_spec_version` to be listed in the checker output.  Checkers using
the former will raise a `DeprecationWarning` indicating the checker is
not preferred.  Also refactors some of the check suite logic for
printing listings and adds unit test for aforementioned behavior.
Addresses #610.