Adds `-D`/`--describe-checks` feature
Adds `-D`/`--describe-checks` feature.  User can specify this flag to
show abridged descriptions of checks implemented by certain checker
suites.  Internally, partial docstrings from functions beginning with the name
`check_` are used to implement this functionality. The behavior depends
upon whether one or more `-t`/`--test` options is present, or if it is
absent completely:

- If `-t`/`--test` is present, all the check functions specified in the
  checkers specified by `-t`/`--test` function are listed, along with
  their descriptions

- If `-t`/`--test` is omitted, implemented checks for all available
  checkers are listed.  Note that the check suites of the form
  `<check_suite>:latest` and `<check_suite>` are omitted, as they refer
  to the latest version of the check suite, and would result in
  redundant information being output.

Useful in combination with `-l` output, or to describe the behavior of
certain checks should a user wish to disable some with
`-s`/`--skip-checks`.