Modification and test fix for check_packed_data()
An annoying string representation conflict between Python 2 and Python 3 was
causing the test_check_packed_data() to fail for the CF check to examine
packed data. This commit now implements using the "name" attribute of the
data types, which is common across Python 2 and Python 3, and makes the strings
unicode so the tests are expecting the correct output.