Adds a new method `get_test_ctx` to reuse `TestCtx`
Adds a new method to `compliance_checker.base.BaseCheck` called
`get_test_ctx`, which attempts to fetch a `TestCtx` from a defaultdict
specific to the checker instance.  If the check does not exist, it is
created in the dict, otherwise, the corresponding dict entry is returned
for further modifications.  Allows predefined `TestCtx` objects to be
reused across functions so that the same `TestCtx` doesn't have to be
passed around one monolithic function.  Potentially also useful for
extending checking behavior defined in newer checker versions, i.e.
additions to check behavior between CF versions.