mod. check of req. vars with std.name for grid map; fixed #691
Modified the checking of the number of variables with spezific
attributes required by a grid mapping. Previously `==1` was checked (are
variable with a specific required standard name was allowed to exist
only once). Now, `>=1` is checked (the variable has to exist at least
once). The user output was updated accordingly.