penalize 'unidentifiable feature for variable'; fixed #679
If 'unidentifiable feature for variable' arose in the past, it was not
penalized meaning that no individual error was thrown. However, this led
to the issue that the message did not appear in the output of compliance
checker if no other error/warning/information was printed out for
section 9.1. Now, it is regularly penalized and a message appears every
time.