generate netCDF4 file from cdl with ncgen
When a CDL file is provided, currently, a netCDF classic file is
created by ncgen. However, a CDL file might contain netCDF4 features
such as attributes/variables of type int64 (indicated by `LL` after
numbers). Therefore, the `ncgen`-call that creates a netCDF file
was modified to create a netCDF4 instead of a netCDF-classic file.
If this does not work, we try to create a netCDF Classic file. If
that does not work, we exit the program.