Adds splitter validation, change attr_check interface
Adds ability to use splitter functions to split a string or other data
type as an argument to a validator constructor.  This allows for
handling comma separated values, for example.  Changes attr_check
interface to use validator instances rather than classes so that custom
splitter functions can be passed in.