Adds ability to fetch latest CF standard name table
Adds an option to the `-d` flag to download the latest CF standard name
table without needing to explicitly specify a version.  This is expected
to be useful in cases with non-interactive runs () of the compliance
checker (e.g. scheduled checker runs) where an up to date standard name
table is desired.