Fetch remote NetCDF resources
Use the content-type request header to inform how a dataset
should be created. Remote NetCDF resources can be created via
fetching the raw bytes and instantiating a MemoizedDataset object.
Extend this capability to fetch TableDAP datasets by simply adding
the .ncCF extension and the string of variables to the dataset URL.